split: errors into trace and legacy parts
