+++ weave/legacy/clear_cache.py
@@ -3,7 +3,6 @@ import time
 
 from weave.legacy import cache
 
-# TODO: This should be split out into a scripts dir
 # Script to run to delete expired caches
 if __name__ == "__main__":
     print("Starting clear cache job", flush=True)
