+++ weave/legacy/trace_legacy.py
@@ -3,8 +3,9 @@ import typing
 
 from weave.legacy import artifact_local, graph, runs
 
-from . import errors, ref_base
-from . import weave_types as types
+from .. import ref_base
+from . import errors
+from .. import weave_types as types
 
 
 def get_obj_creator(ref: ref_base.Ref) -> typing.Optional[runs.Run]:
